removed an unnecessary cast

+++ crates/bevy_ui/src/render/mod.rs
@@ -645,7 +645,7 @@ pub fn extract_text_uinodes(
         // * Round the physical position to the nearest physical pixel
         // * Multiply by the rounded physical position by the inverse scale factor to return to logical coordinates
         let logical_top_left = translation.xy() - 0.5 * uinode.size();
-        let physical_nearest_pixel = (logical_top_left * scale_factor as f32).round();
+        let physical_nearest_pixel = (logical_top_left * scale_factor).round();
         let logical_top_left_nearest_pixel = physical_nearest_pixel * inverse_scale_factor;
         let transform = Mat4::from_scale_rotation_translation(
             scale,
